requirements.txt
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
anyio==4.9.0
astor==0.8.1
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
Cython==3.0.12
decorator==5.2.1
fire==0.7.0
fonttools==4.56.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.37.0
lazy_loader==0.4
lmdb==1.6.2
lxml==5.3.1
networkx==3.4.2
numpy==2.2.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
packaging==24.2
paddleocr==2.10.0
paddlepaddle==3.0.0b2
pillow==11.1.0
protobuf==6.30.1
pyclipper==1.3.0.post6
pydantic==2.11.1
pydantic_core==2.33.0
python-docx==1.1.2
PyYAML==6.0.2
RapidFuzz==3.12.2
requests==2.32.3
scikit-image==0.25.2
scipy==1.15.2
shapely==2.0.7
simsimd==6.2.1
sniffio==1.3.1
soupsieve==2.6
stringzilla==3.12.3
termcolor==2.5.0
tifffile==2025.3.13
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.12.2
urllib3==2.3.0
